Updated Plugins
2 Click Social Media Buttons 0.32.2
BuddyPress 1.5.5
Contact Form 7 3.1.1
Events Manager 5.1.5
GD Star Rating 1.9.16
Really Simple CAPTCHA 1.5
User Role Editor 3.5.4

+++ wp-content/plugins/events-manager/classes/em-event-post-admin.php
@@ -40,7 +40,7 @@ class EM_Event_Post_Admin{
 			} elseif ( $EM_Event->is_recurrence() ) {
 				$warning = "<p><strong>".__('WARNING: This is a recurrence in a set of recurring events.', 'dbem')."</strong></p>";
 				$warning .= "<p>". sprintf(__('If you update this event data and save, it could get overwritten if you edit the recurring event template. To make it an independent, <a href="%s">detach it</a>.', 'dbem' ), $EM_Event->get_detach_url())."</p>";
-				$warning .= "<p>".sprintf(__('To manage the whole set, <a href="%s">edit the recurring event template</a>.', 'dbem'),admin_url().'post.php?action=edit&amp;post='.$EM_Event->recurrence_id)."</p>";
+				$warning .= "<p>".sprintf(__('To manage the whole set, <a href="%s">edit the recurring event template</a>.', 'dbem'),admin_url('post.php?action=edit&amp;post='.$EM_Event->get_event_recurrence()->post_id))."</p>";
 				?><div class="updated"><?php echo $warning; ?></div><?php
 			}
 			if( !empty($EM_Event->group_id) && function_exists('groups_get_group') ){
@@ -70,8 +70,8 @@ class EM_Event_Post_Admin{
 		if(!defined('UNTRASHING_'.$post_id) && $is_post_type && $saving_status ){
 			if( wp_verify_nonce($_REQUEST['_emnonce'], 'edit_event') ){ 
 				//this is only run if we know form data was submitted, hence the nonce
-				do_action('em_event_save_pre', $EM_Event); //technically, the event is saved... but the meta isn't. wp doesn't give an pre-intervention action for this (or does it?)
 				$EM_Event = em_get_event($post_id, 'post_id');
+				do_action('em_event_save_pre', $EM_Event); //technically, the event is saved... but the meta isn't. wp doesn't give an pre-intervention action for this (or does it?)
 				//Handle Errors by making post draft
 				$get_meta = $EM_Event->get_post_meta();
 				$save_meta = $EM_Event->save_meta();
@@ -99,6 +99,7 @@ class EM_Event_Post_Admin{
 				//we're updating only the quick-edit style information, which is only post info saved into the index
 				$EM_Event = em_get_event($post_id, 'post_id'); //grab event, via post info
 				if( $EM_Event->validate() ){
+					do_action('em_event_save_pre', $EM_Event); //technically, the event is saved... but the meta isn't. wp doesn't give an pre-intervention action for this (or does it?)
 					//first things first, we must make sure we have an index, if not, reset it to a new one:
 					$event_truly_exists = $wpdb->get_var('SELECT event_id FROM '.EM_EVENTS_TABLE." WHERE event_id={$EM_Event->event_id}") == $EM_Event->event_id;
 					if(empty($EM_Event->event_id) || !$event_truly_exists){ $EM_Event->save_meta(); }
@@ -119,6 +120,7 @@ class EM_Event_Post_Admin{
 					}
 					apply_filters('em_event_save', true, $EM_Event);
 				}else{
+					do_action('em_event_save_pre', $EM_Event); //technically, the event is saved... but the meta isn't. wp doesn't give an pre-intervention action for this (or does it?)
 					//Event doesn't validate, so set status to null
 					$EM_Event->set_status(null, true);
 					apply_filters('em_event_save', false, $EM_Event);
@@ -191,9 +193,6 @@ class EM_Event_Post_Admin{
 		if( get_option('dbem_attributes_enabled', true) ){
 			add_meta_box('em-event-attributes', __('Attributes','dbem'), array('EM_Event_Post_Admin','meta_box_attributes'),EM_POST_TYPE_EVENT, 'normal','default');
 		}
-		if( (empty($EM_Event->event_id) || !empty($EM_Event->group_id)) && function_exists('groups_get_user_groups') ){
-			add_meta_box('em-event-group', __('Group Ownership','dbem'), array('EM_Event_Post_Admin','meta_box_group'),EM_POST_TYPE_EVENT, 'side','low');
-		}
 		if( EM_MS_GLOBAL && !is_main_site() && get_option('dbem_categories_enabled') ){
 			add_meta_box('em-event-categories', __('Site Categories','dbem'), array('EM_Event_Post_Admin','meta_box_ms_categories'),EM_POST_TYPE_EVENT, 'side','low');
 		}
@@ -225,10 +224,6 @@ class EM_Event_Post_Admin{
 	
 	function meta_box_attributes(){
 		em_locate_template('forms/event/attributes.php',true);
-	}
-	
-	function meta_box_group(){
-		em_locate_template('forms/event/group.php',true);
 	}
 	
 	function meta_box_location(){
@@ -300,7 +295,7 @@ class EM_Event_Recurring_Post_Admin{
 			$events_array = EM_Events::get( array('recurrence'=>$EM_Event->event_id, 'scope'=>'all', 'status'=>'all' ) );
 			foreach($events_array as $event){
 				/* @var $event EM_Event */
-				if($EM_Event->event_id == $event->recurrence_id ){ //double check the event is a recurrence of this event
+				if($EM_Event->event_id == $event->recurrence_id && !empty($event->recurrence_id) ){ //double check the event is a recurrence of this event
 					wp_delete_post($event->post_id, true);
 				}
 			}
@@ -366,6 +361,9 @@ class EM_Event_Recurring_Post_Admin{
 		}
 		if( (empty($EM_Event->event_id) || !empty($EM_Event->group_id)) && function_exists('groups_get_user_groups') ){
 			add_meta_box('em-event-group', __('Group Ownership','dbem'), array('EM_Event_Post_Admin','meta_box_group'),'event-recurring', 'side','low');
+		}
+		if( EM_MS_GLOBAL && !is_main_site() && get_option('dbem_categories_enabled') ){
+			add_meta_box('em-event-categories', __('Site Categories','dbem'), array('EM_Event_Post_Admin','meta_box_ms_categories'),'event-recurring', 'side','low');
 		}
 	}
